+++ guard/src/commands/reporters/mod.rs
@@ -1,3 +1,6 @@
+pub mod test;
+pub mod validate;
+
 use std::{fmt::Display, rc::Rc, time::Instant};
 
 use quick_xml::{
@@ -7,14 +10,11 @@ use quick_xml::{
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    commands::{
-        validate::{structured::StructuredReporter, DataFile},
-        ERROR_STATUS_CODE, FAILURE_STATUS_CODE,
-    },
+    commands::{validate::DataFile, ERROR_STATUS_CODE, FAILURE_STATUS_CODE},
     rules::{
         self,
         eval::eval_rules_file,
-        eval_context::{root_scope, simplifed_json_from_root, FileReport, Messages},
+        eval_context::{root_scope, simplifed_json_from_root, Messages},
         exprs::RulesFile,
         Status,
     },
@@ -34,7 +34,7 @@ impl<'report> JunitReport<'report> {
         &self,
         writer: &'report mut crate::utils::writer::Writer,
     ) -> crate::rules::Result<()> {
-        let mut writer = Writer::new_with_indent(writer, b' ', 4);
+        let mut writer = quick_xml::Writer::new_with_indent(writer, b' ', 4);
         let decl = BytesDecl::new("1.0", Some("UTF-8"), None);
 
         writer.write_event(Event::Decl(decl))?;
@@ -67,75 +67,6 @@ impl<'reporter> JunitReporter<'reporter> {
         {
             self.exit_code = code;
         }
-    }
-}
-
-impl<'reporter> StructuredReporter for JunitReporter<'reporter> {
-    fn report(&mut self) -> rules::Result<i32> {
-        let now = Instant::now();
-        let mut suites = vec![];
-        let mut total_errors = 0;
-        let mut total_failures = 0;
-        let mut tests = 0;
-
-        for each in &self.data {
-            let file_report = FileReport {
-                name: &each.name,
-                ..Default::default()
-            };
-
-            let mut failures = 0;
-            let mut errors = 0;
-
-            let test_cases = self.rules.iter().try_fold(
-                vec![],
-                |mut test_cases, (rule, name)| -> rules::Result<Vec<TestCase<'_>>> {
-                    let tc = get_test_case(each, rule, name)?;
-
-                    if matches!(tc.status, TestCaseStatus::Fail(_)) {
-                        failures += 1;
-                    } else if matches!(tc.status, TestCaseStatus::Error { .. }) {
-                        errors += 1;
-                    }
-
-                    tests += 1;
-                    test_cases.push(tc);
-                    Ok(test_cases)
-                },
-            )?;
-
-            let suite = TestSuite {
-                name: file_report.name.to_string(),
-                test_cases,
-                time: now.elapsed().as_millis(),
-                errors,
-                failures,
-            };
-
-            total_errors += errors;
-            total_failures += failures;
-
-            suites.push(suite);
-        }
-
-        if total_errors > 0 {
-            self.update_exit_code(ERROR_STATUS_CODE)
-        } else if total_failures > 0 {
-            self.update_exit_code(FAILURE_STATUS_CODE)
-        }
-
-        let report = JunitReport {
-            name: "cfn-guard validate report",
-            test_suites: suites,
-            failures: total_failures,
-            errors: total_errors,
-            tests,
-            duration: now.elapsed().as_millis(),
-        };
-
-        report.serialize(self.writer)?;
-
-        Ok(self.exit_code)
     }
 }
 
@@ -212,7 +143,6 @@ pub struct TestCase<'test> {
     pub time: u128,
     pub(crate) status: TestCaseStatus,
 }
-
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub(crate) enum TestCaseStatus {
     Pass,
@@ -220,7 +150,6 @@ pub(crate) enum TestCaseStatus {
     Fail(FailingTestCase),
     Error { error: String },
 }
-
 #[derive(Debug, Clone)]
 pub struct TestSuite<'suite> {
     pub name: String,
@@ -229,19 +158,33 @@ pub struct TestSuite<'suite> {
     pub errors: usize,
     pub failures: usize,
 }
-
+impl<'suite> TestSuite<'suite> {
+    pub fn new(
+        name: String,
+        test_cases: Vec<TestCase<'suite>>,
+        time: u128,
+        errors: usize,
+        failures: usize,
+    ) -> Self {
+        Self {
+            name,
+            test_cases,
+            time,
+            errors,
+            failures,
+        }
+    }
+}
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub(crate) struct FailingTestCase {
     pub(crate) name: Option<String>,
     pub(crate) messages: Vec<Messages>,
 }
-
 #[derive(Default, Debug)]
 struct Failure<'report> {
     name: Option<&'report String>,
     messages: Vec<&'report String>,
 }
-
 #[derive(Default, Debug)]
 pub struct TestSuites<'report, 'se: 'report> {
     pub name: &'report str,
@@ -251,7 +194,6 @@ pub struct TestSuites<'report, 'se: 'report> {
     pub time: u128,
     pub test_suites: &'se [TestSuite<'report>],
 }
-
 #[derive(Debug)]
 enum EventType<'report, 'se: 'report> {
     Failure(Failure<'report>),
@@ -260,7 +202,6 @@ enum EventType<'report, 'se: 'report> {
     TestSuite(&'se TestSuite<'report>),
     TestSuites(TestSuites<'report, 'se>),
 }
-
 impl<'report, 'se: 'report> EventType<'report, 'se> {
     fn serialize_start_event(
         &self,
@@ -269,18 +210,15 @@ impl<'report, 'se: 'report> EventType<'report, 'se> {
     ) -> crate::rules::Result<()> {
         Ok(writer.write_event(Event::Start(tag))?)
     }
-
     fn start_tag(&self) -> BytesStart<'_> {
         BytesStart::new(self.to_string())
     }
-
     fn serialize_end_event(
         &self,
         writer: &mut Writer<impl std::io::Write>,
     ) -> crate::rules::Result<()> {
         Ok(writer.write_event(Event::End(BytesEnd::new(self.to_string())))?)
     }
-
     fn extend_attributes(&self, tag: &mut BytesStart<'_>) {
         match self {
             EventType::Failure(failure) => {
@@ -288,17 +226,14 @@ impl<'report, 'se: 'report> EventType<'report, 'se> {
                     tag.push_attribute(("message", name.as_str()));
                 }
             }
-
             EventType::TestCase(test_case) => {
                 if let Some(id) = test_case.id {
                     tag.push_attribute(("id", id));
                 }
-
                 tag.extend_attributes([
                     ("name", test_case.name),
                     ("time", format!("{:.3}", test_case.time).as_str()),
                 ]);
-
                 match &test_case.status {
                     TestCaseStatus::Fail(..) => {}
                     status => {
@@ -308,7 +243,6 @@ impl<'report, 'se: 'report> EventType<'report, 'se> {
                             TestCaseStatus::Error { .. } => "error",
                             _ => unreachable!(),
                         };
-
                         tag.extend_attributes([("status", status)]);
                     }
                 }
@@ -333,11 +267,9 @@ impl<'report, 'se: 'report> EventType<'report, 'se> {
             }
         }
     }
-
     fn serialize(&self, writer: &mut Writer<impl std::io::Write>) -> crate::rules::Result<()> {
         let mut tag = self.start_tag();
         self.extend_attributes(&mut tag);
-
         match self {
             EventType::Failure(failure) => {
                 if !failure.messages.is_empty() {
@@ -352,21 +284,17 @@ impl<'report, 'se: 'report> EventType<'report, 'se> {
                 TestCaseStatus::Fail(failure) => {
                     self.serialize_start_event(writer, tag)?;
                     let name = failure.name.as_ref();
-
                     let event = match failure.messages.is_empty() {
                         false => {
                             let messages = failure.messages.iter().fold(vec![], |mut acc, msg| {
                                 if let Some(custom_message) = &msg.custom_message {
                                     acc.push(custom_message);
                                 }
-
                                 if let Some(error_message) = &msg.error_message {
                                     acc.push(error_message);
                                 }
-
                                 acc
                             });
-
                             EventType::Failure(Failure { name, messages })
                         }
                         true => EventType::Failure(Failure {
@@ -374,7 +302,6 @@ impl<'report, 'se: 'report> EventType<'report, 'se> {
                             messages: vec![],
                         }),
                     };
-
                     event.serialize(writer)?;
                     self.serialize_end_event(writer)?;
                 }
