+++ guard/src/commands/reporters/validate/common.rs
