+++ guard/src/commands/reporters/validate/console_reporter.rs
