+++ guard/src/commands/reporters/validate/generic_summary.rs
